Preliminary fix to issue #11.  It works by splitting up the field by dots.  For example, filtering on: input,report.body_proportion
Will filter on the "input" field, which is only one level deep, but it will also recurse into the body_proportion subfield under "report".

This is a little problematic, because if I want to filter on the HTTP return codes in http_requests, I need something like:
report.requests.0.reponse.code

And in the current setup, the user has no way of knowing there's a list there.  It might make sense to refactor
LoadRecursiveCollection so that it notes list indices instead of just spitting out "nested data"

PYYAML Y U NO MAKE EVERYTHING A DICTIONARY??!!